requirements.txt
# GUI
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2

# OCR
pytesseract==0.3.10
Pillow==10.0.0

# 图像处理
opencv-python==4.8.0.74
numpy==1.24.3

# 系统交互
pyautogui==0.9.54
pynput==1.7.6

# 工具库
python-dotenv==1.0.0
loguru==0.7.0

# 测试
pytest==7.4.0
